map works, need live projects and remove logs
