+++ src/css/custom.css
@@ -537,13 +537,7 @@ a:hover {
 /*************/
 /* Home Page */
 /*************/
-#home-map-pane {
-    width: 75%;
-    position: fixed;
-}
-#home-map-pane .ol-attribution {
-    position: fixed;
-}
+
 #image-analysis-pane .ol-attribution {
     right: 10vw;
 }
