+++ src/css/highlights.css
@@ -24,11 +24,13 @@
     align-items: flex-start;
     gap: 4px;
 }
+
 .header-link {
     display: flex;
     align-items: center;
     gap: 4px;
 }
+
 .header-link span {
     color: var(--Primary-Standard-Green, #1F7067);
     text-align: justify;
@@ -286,19 +288,22 @@
     display: flex;
     height: 876px;
     align-items: flex-start;
-    gap: 16px;}
+    gap: 16px;
+    width: 100%;}
 
 #projects-map-container {
     display: flex;
-    width: 480px;
     flex-direction: column;
     align-items: flex-start;
     gap: 16px;
+    flex-grow: 1;
 }
 
 #projects-map {
+    position: relative;
+    width: 100%;
     display: flex;
-    height: 860px;
+
     padding-top: 1px;
     flex-direction: column;
     justify-content: center;
@@ -461,22 +466,22 @@
 }
 
 .expand-description {
-display: flex;
-align-items: center;
-gap: 8px;
-cursor: pointer;
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    cursor: pointer;
 }
 
 .expand-description span {
-color: var(--Primary-Standard-Green, #1F7067);
-font-feature-settings: 'liga' off, 'clig' off;
+    color: var(--Primary-Standard-Green, #1F7067);
+    font-feature-settings: 'liga' off, 'clig' off;
 
-/* Link/Default */
-font-family: "Public Sans";
-font-size: 14px;
-font-style: normal;
-font-weight: 500;
-line-height: normal;}
+    /* Link/Default */
+    font-family: "Public Sans";
+    font-size: 14px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: normal;}
 
 .project-controls {
     display: flex;
@@ -485,3 +490,13 @@ line-height: normal;}
 }
 
 
+#home-map-pane {
+    width: 100%;
+    position: absolute;
+}
+#home-map-pane .ol-attribution {
+    position: fixed;
+}
+
+#mapPanel {
+    flex-grow: 1;}
